Graph Fix
Fixed an issue where the modal graph and display graphs shared an object in memory when both were in the same family (e.g., both barometer). This resulted in some weird issues.

Also adjusted it so that the selected modal graph is persistent (e.g., stays the same when closing and opening the modal).